Adds /tags/ API
